+++ src/main/datatypes/MQ2DataTypes.h
@@ -754,7 +754,21 @@ public:
 };
 
 //============================================================================
-// MQ2InvSlotWindowType
+// MQHotButtonWindowType
+
+class MQHotButtonType : public MQ2Type
+{
+public:
+	MQHotButtonType();
+
+	bool GetMember(MQVarPtr VarPtr, const char* Member, char* Index, MQTypeVar& Dest) override;
+	bool ToString(MQVarPtr VarPtr, char* Destination) override;
+
+	bool FromData(MQVarPtr& VarPtr, const MQTypeVar& Source) override;
+};
+
+//============================================================================
+// MQInvSlotWindowType
 
 class MQInvSlotWindowType : public MQ2Type
 {
@@ -1504,6 +1518,7 @@ public:
 	MQSocialType();
 
 	bool GetMember(MQVarPtr VarPtr, const char* Member, char* Index, MQTypeVar& Dest) override;
+	bool ToString(MQVarPtr VarPtr, char* Destination) override;
 
 	static bool dataSocial(const char* szIndex, MQTypeVar& Ret);
 };