+++ data/resources/CHANGELOG.md
@@ -1,3 +1,7 @@
+July 7, 2024:
+- Add new HotButton window type: ${Window[HotButtonWnd2/HB_Button6].HotButton}
+- See docs at: https://docs.macroquest.org/reference/data-types/datatype-hotbuttonwindow/
+
 July 3, 2024:
 - Fix EverQuest.Ping, add EverQuest.ConnectionStrength - these are the value from the net meter.
 - Fix UI not working after persona swap (#797).